+++ pyproject.toml
@@ -1,0 +1,23 @@
+[tool.poetry]
+name = "lunch-web"
+version = "0.1.0"
+description = "REST API for parsing menus from restaurants"
+authors = ["Pavel Yadlouski <pavel.yadlouski@gmail.com>"]
+readme = "README.md"
+packages = [{include = "lunch_web"}]
+
+[tool.poetry.dependencies]
+python = "^3.10"
+Flask = "^2.2.2"
+Flask-RESTful = "^0.3.9"
+requests = "^2.28.1"
+gunicorn = "^20.1.0"
+beautifulsoup4 = "^4.11.1"
+python-dateutil = "^2.8.2"
+
+[tool.poetry.group.dev.dependencies]
+pytest = "^7.2.0"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
