--- requirements.txt
@@ -1,8 +1,0 @@
-gunicorn>=20
-python-dotenv>=0.20.0
-requests>=2.28
-python-dateutil>=2.8
-Flask>=2.2
-Flask-RESTful>=0.3
-beautifulsoup4>=4.11.1
-pytest>=7.2.0
